replace dd with own implementation for creating a file with size
